+++ src/ui.py
@@ -4,7 +4,7 @@
 from __future__ import (unicode_literals, division, absolute_import, print_function)
 
 __license__ = 'GPL v3'
-__copyright__ = '2024, Your Name'
+__copyright__ = '2024, LeviTK'
 __docformat__ = 'restructuredtext en'
 
 import os